Library unida

+++ Library/Ventas.cs
@@ -8,19 +8,19 @@ using System.Threading.Tasks;
 
 namespace Library
 {
-	public class Ventas
-	{
+    public class Ventas
+    {
 
-		[Key]
-		public int VentaId { get; set; }
+        [Key]
+        public int VentaId { get; set; }
 
-		[Required(ErrorMessage = "Este campo es requerido")]
-		public DateTime Fecha { get; set; } = DateTime.Today;
+        [Required(ErrorMessage = "Este campo es requerido")]
+        public DateTime Fecha { get; set; } = DateTime.Today;
 
-		[Range(1, int.MaxValue, ErrorMessage = "El campo debe ser mayor que cero")]
-		public decimal Valor { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "El campo debe ser mayor que cero")]
+        public float Valor { get; set; }
 
-		[ForeignKey("VentaId")]
-		public List<VentaDetalle> VentaDetalle { get; set; } = new List<VentaDetalle>();
-	}
+        [ForeignKey("VentaId")]
+        public List<VentaDetalle> VentaDetalle { get; set; } = new List<VentaDetalle>();
+    }
 }
